requirements.txt
APScheduler==3.10.4
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cpf-generator==2.1.4
cryptography==43.0.0
deprecation==2.1.0
exceptiongroup==1.2.2
Flask==3.0.3
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
pycparser==2.22
pytest==8.3.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
starkbank==2.26.0
starkbank-ecdsa==2.2.0
starkcore==0.5.0
tomli==2.0.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
Werkzeug==3.0.3
